+++ pyproject.toml
@@ -8,7 +8,9 @@ requires-python = ">=3.8"
 dependencies = [
     "langchain==0.1.1",
     "SQLAlchemy==2.0.7",
-    "cloud-sql-python-connector[asyncpg]==1.5.0"
+    "cloud-sql-python-connector[asyncpg]==1.5.0",
+    "nest_asyncio",
+    "pgvector"
 ]
 
 [project.urls]
