+++ src/langchain_google_cloud_sql_pg/__init__.py
@@ -11,3 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from langchain_google_cloud_sql_pg.postgresql_engine import PostgreSQLEngine
+from langchain_google_cloud_sql_pg.vectorstore import CloudSQLVectorStore
+
+__all__ = ["PostgreSQLEngine", "CloudSQLVectorStore"]
